Big Commitments.
Started with the list view. Putting edit aside, so the functions to view edit is going to be committed later. Also, edit will also suffice as adding a new ToDo. The function for edit will request a ToDoObject, but it is nullable. When null, it will assume entering a new object.